backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
